fix close tun device

+++ world/bkeka/vpn_tools.py
@@ -24,7 +24,7 @@ SUCCESS_MESSAGE = "Initialization Sequence Completed"
 OPENVPN_COMMAND = "openvpn"
 VPN_SUCCESS = 1
 VPN_ERROR = 0
-MAX_RETRIES = 5
+MAX_RETRIES = 255
 
 # Global variable for VPN subprocess
 openvpn_process = None
@@ -78,13 +78,9 @@ def get_config_file(vpn_config_dir):
 def close_vpn_connection():
     global openvpn_process
     # kill() does not always closes tun0 device
-    if openvpn_process is not None:
+    if openvpn_process is not None and openvpn_process.poll() is not None:
         openvpn_process.kill()
         openvpn_process = None
-    # 'sudo killall openvpn' does something nasty to VPN provider connection
-    process = Popen(['sudo', 'killall', OPENVPN_COMMAND], stdout=PIPE, stderr=PIPE, shell=False)
-    if not pipe_no_wait(process):
-        raise OpenVPNProcessExceptions("Failed to set file descriptor non blocking")
 
 
 def connect_to_vpn(vpn_config_dir, vpn_credentials_file):
